select completed models
