+++ application/ui/src/features/inspect/models/models-view.component.tsx
@@ -1,7 +1,6 @@
 import { useMemo } from 'react';
 
-import { $api } from '@geti-inspect/api';
-import { usePipeline, useProjectIdentifier } from '@geti-inspect/hooks';
+import { usePipeline } from '@geti-inspect/hooks';
 import {
     Cell,
     Column,
@@ -18,23 +17,15 @@ import {
 import { sortBy } from 'lodash-es';
 import { useDateFormatter } from 'react-aria';
 
-import { useProjectTrainingJobs, useRefreshModelsOnJobUpdates } from '../dataset/dataset-status-panel.component';
+import { useCompletedModels } from '../../../hooks/use-completed-models.hook';
+import { useProjectTrainingJobs } from '../../../hooks/use-project-trainingJobs.hook';
+import type { ModelData } from '../../../hooks/utils';
+import { useRefreshModelsOnJobUpdates } from '../dataset/dataset-status-panel.component';
 import { formatSize } from '../utils';
 import { ModelActionsMenu } from './model-actions-menu.component';
 import { ModelStatusBadges } from './model-status-badges.component';
-import { ModelData } from './model-types';
 
 import classes from './models-view.module.scss';
-
-const useModels = () => {
-    const { projectId } = useProjectIdentifier();
-    const modelsQuery = $api.useSuspenseQuery('get', '/api/projects/{project_id}/models', {
-        params: { path: { project_id: projectId } },
-    });
-    const models = modelsQuery.data.models;
-
-    return models;
-};
 
 export const ModelsView = () => {
     const { data: pipeline } = usePipeline();
@@ -42,39 +33,9 @@ export const ModelsView = () => {
 
     const dateFormatter = useDateFormatter({ dateStyle: 'medium', timeStyle: 'short' });
     const selectedModelId = pipeline.model?.id;
+    const models = useCompletedModels();
+
     useRefreshModelsOnJobUpdates(jobs);
-
-    const models = useModels()
-        .filter((model) => model.is_ready)
-        .map((model): ModelData | null => {
-            const job = jobs.find(({ id }) => id === model.train_job_id);
-            if (job === undefined) {
-                return null;
-            }
-
-            let timestamp = '';
-            let durationInSeconds = 0;
-            const start = job.start_time ? new Date(job.start_time) : new Date();
-            if (job) {
-                const end = job.end_time ? new Date(job.end_time) : new Date();
-                durationInSeconds = Math.floor((end.getTime() - start.getTime()) / 1000);
-                timestamp = dateFormatter.format(start);
-            }
-
-            return {
-                id: model.id!,
-                name: model.name!,
-                status: 'Completed',
-                architecture: model.name!,
-                startTime: start.getTime(),
-                timestamp,
-                durationInSeconds,
-                progress: 1.0,
-                job,
-                sizeBytes: model.size ?? null,
-            };
-        })
-        .filter((model): model is ModelData => model !== null);
 
     const completedModelsJobsIDs = new Set(models.map((model) => model.job?.id));
