+++ application/ui/src/hooks/utils.ts
